Camera/requirements.txt
cryptography==3.4.6
face-recognition==1.3.0
ffmpeg-python==0.2.0
Flask==1.1.2
PyJWT==2.0.1
python-dotenv==0.15.0
requests==2.25.1
tinydb==4.4.0
gunicorn==20.1.0